+++ src/components/Common/Button/Button.module.css
@@ -1,6 +1,7 @@
 .button {
     height: 4.8rem;
 
+    margin: 0 auto;
     padding: 1.2rem 2rem;
 
     font-size: 1.6rem;
