+++ src/components/Modal/ConfirmModal/ConfirmModal.module.css
@@ -20,59 +20,15 @@
     line-height: 1.79rem;
 }
 
-.complete {
-    height: 4.8rem;
-
-    margin: 0 auto;
-    padding: 1.3rem 4.8rem;
-    border-radius: 5rem;
-    border: none;
-
-    color: white;
-    font-size: 1.6rem;
-    font-weight: 600;
-    line-height: 1.909rem;
-
-    background-color: var(--primary-orange);
-
-    cursor: pointer;
-}
-
 .buttons {
     display: flex;
-    justify-content: space-between;
+    justify-content: center;
+
+    width: 100%;
 
     margin: 0 auto;
 
     gap: 1rem;
-}
-
-.buttons button {
-    height: 4.8rem;
-
-    padding: 1.3rem 4.8rem;
-
-    font-size: 1.6rem;
-    font-weight: 600;
-    line-height: 1.909rem;
-
-    cursor: pointer;
-}
-
-.cancel {
-    border-radius: 5rem;
-    border: 0.1rem solid var(--primary-orange);
-
-    color: var(--primary-orange);
-    background-color: transparent;
-}
-
-.confirm {
-    border-radius: 5rem;
-    border: none;
-
-    color: white;
-    background-color: var(--primary-orange);
 }
 
 /***** Mobile *****/
@@ -82,20 +38,6 @@
     }
 
     .content span {
-        font-size: 1.4rem;
-        line-height: 1.671rem;
-    }
-
-    .complete {
-        height: 4rem;
-
-        font-size: 1.4rem;
-        line-height: 1.671rem;
-    }
-
-    .buttons button {
-        height: 4rem;
-
         font-size: 1.4rem;
         line-height: 1.671rem;
     }
