+++ src/components/StartupDetail/StartupDetailInvest.module.css
@@ -18,6 +18,10 @@
   height: 100%;
 }
 
+.header button {
+  margin: 0;
+}
+
 .header h1 {
   font-size: 2rem;
   font-weight: 700;
